Reparameterization of curves by arc length.
- Works for parametric-point, spline-curve, and plane-curve curves.
- Some issues with plane curves:

For plane-curves, there are some hacks to get around the chain rule
or integral substitution in calculus operations, because of the
requirement of using a normalized domain [0,1] and lerping it to the
domains of the component functions. Currently, arc length
reparameterization will not work on plane-curves with component
functions that are mathematically equivalent to #'* (identity), but are
not literally #'* as the component function.

Potential fixes:

- This could probably be heuristically checked with a few derivatives
over the domain.
- Restrict possible input functions. This is probably a good idea
anyway.
- Or just use automatic differentiation and it will always be correct.